+++ src/main/java/core/InputHandler.java
@@ -13,6 +13,7 @@ import core.exceptions.ToDoEmptyException;
 import core.tasks.Deadline;
 import core.tasks.Event;
 import core.tasks.Task;
+import core.tasks.TaskList;
 import core.tasks.ToDo;
 import utilities.OutputFormatter;
